// dropbox/_generated/server.d.ts
type GenericCtx =
  | GenericActionCtx<DataModel>
  | GenericMutationCtxWithTable<DataModel>
  | GenericQueryCtxWithTable<DataModel>;

/**
 * Define a query in this Convex app's public API.
 *
 * This function will be allowed to read your Convex database and will be accessible from the client.
 *
 * @param func - The query function. It receives a {@link QueryCtx} as its first argument.
 * @returns The wrapped query. Include this as an `export` to name it and make it accessible.
 */
export declare const query: QueryBuilder<DataModel, "public">;
// ...
export declare const internalQuery: QueryBuilder<DataModel, "internal">;

/**
 * Define a mutation in this Convex app's public API.
 *
 * This function will be allowed to modify your Convex database and will be accessible from the client.
 *
 * @param func - The mutation function. It receives a {@link MutationCtx} as its first argument.
 * @returns The wrapped mutation. Include this as an `export` to name it and make it accessible.
 */
export declare const mutation: MutationBuilder<DataModel, "public">;
// ...
export declare const action: ActionBuilder<DataModel, "public">;
// ...
export declare const component: {
  oauth: {
    getAccessToken: FunctionReference<
      "query",
      "internal",
      { identity: string },
      string
    >;
    isLinked: FunctionReference<"query", "internal", { identity: string }, any>;
    link: FunctionReference<
      "mutation",
      "internal",
      { csrfToken: string; identity?: string; makeUrl: string },
      { identity: string; url: string }
    >;
    setCsrfUrl: FunctionReference<"query", "internal", {}, string>;
  };
};
type ComponentArgs = {
  CONVEX_SITE: string;
  appKey: string;
  appSecret: string;
  redirectAfterLogin: string;
};
export declare const componentArg: <Name extends keyof ComponentArgs>(
  ctx: GenericCtx,
  name: Name,
) => ComponentArgs[Name];